Cache LSB info
If the user really insists on changing the distribution
*while* running this tool, there is no sane behavior
anyway.

This change speeds up the process by several orders of magnitude.
For me personally, this change reduces the running time
from roughly 20 minutes to roughly 5 seconds.
This is because the lsb_release.get_distro_information call
used to create a new process for each package, which is
completely unnecessary.