test_clone_upgrade: avoid false positives when a new release starts
The tests exercise roughtly these steps:
1. create an env of the -2 release (groovy in this case)
2. stores a package selection list of OLD in tarball
3. create an env of the -1 release (hirsute in this case)
4. create an apt-cache of the NEW env
5. restores package selection to the state of "OLD" from tarball
6. expect that there wll be upgrades

But early on in a development cycle the delta will be zero and the test will
fail. That is the reason for bug 1902830 and bug 1510060.

We can make sure we are much more likely to find upgrades but stil exercise
clone the same way by comparing "earliest-supportd vs dev" instead of
"dev vs dev-1".